+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and starting with version v0.1.0, this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Planned]
+- add `--model=Users\User --keys=all,active` options to `artisan make:loader` command.
+
 ## [Unreleased]
-- `php artisan make:loader --model=Users\User --keys=all,active` Generate ModuleLoader stubs, preloaded with empty methods & correct Model
+- `php artisan make:loader` Generate empty Module Loader stubs
 - `Vuex::dd()` dd's the current store data
 
 ## [0.1.0] - 2020-02-13